+++ server/fixtures/locations.json
@@ -1,0 +1,18 @@
+[
+  {
+    "model": "Location",
+    "data": {
+      "address": "Cambridge, MA, USA",
+      "latitude": 42.3736,
+      "longitude": 71.1106
+    }
+  },
+  {
+    "model": "Location",
+    "data": {
+      "address": "San Francisco, CA, USA",
+      "latitude": 37.7833,
+      "longitude": 122.4167
+    }
+  }
+]
